requirements.txt
logzero>=1.7.0
PyYAML>=6.0
requests>=2.26.0